set seed, added train val tst
