+++ run.py
@@ -2,6 +2,7 @@
 import numpy as np
 import matplotlib.pyplot  as plt
 import torch
+import os
 from torch.autograd import Variable
 from torchvision import transforms
 from torch.utils.data import DataLoader
@@ -19,17 +20,6 @@ from process_data.import_test import *
 if __name__ ==  '__main__':
 
     device=torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-    def seed_torch(seed=0):
-        np.random.seed(seed)
-        os.environ['PYTHONHASHSEED'] = str(seed)
-        np.random.seed(seed)
-        torch.manual_seed(seed)
-        torch.cuda.manual_seed(seed)
-        torch.cuda.manual_seed_all(seed) # if you are using multi-GPU.
-        torch.backends.cudnn.enabled = False 
-        torch.backends.cudnn.benchmark = False
-        torch.backends.cudnn.deterministic = True
-        
     seed_torch()
 
     def _init_fn():
